requirements.txt
# MySite 项目依赖包
# 基于Flask的用户管理系统和学术资源库

# ==================== 核心依赖 ====================

# Web框架
Flask>=3.0.0

# 数据库连接
mysql-connector-python>=8.0.0

# 文件上传和表单处理
Werkzeug>=3.0.0

# ==================== 学术资源库依赖 ====================

# 文件处理
Pillow>=10.0.0          # 图片处理和文件格式支持
python-magic>=0.4.27    # 文件类型检测

# 内容处理
markdown>=3.5.0          # Markdown支持
bleach>=6.0.0            # HTML清理和安全

# 搜索功能
whoosh>=2.7.4           # 全文搜索引擎

# ==================== 可选依赖 ====================

# 数据库相关
# PyMySQL>=1.0.0         # 替代MySQL连接器
# SQLAlchemy>=2.0.0      # ORM支持

# 环境管理
# python-dotenv>=1.0.0   # 环境变量管理

# 日志和调试
# colorlog>=6.0.0        # 彩色日志输出

# 数学公式支持
# KaTeX>=0.16.0          # 数学公式渲染

# 开发工具
# pytest>=7.0.0          # 单元测试
# black>=23.0.0          # 代码格式化
# flake8>=6.0.0          # 代码检查
# mypy>=1.0.0            # 类型检查

# ==================== 系统要求 ====================
# Python 3.7+
# MySQL 5.7+ 或 MariaDB 10.3+
# 现代浏览器（支持ES6+）
